try again for 64 bits

+++ CoinUtils/src/CoinMessageHandler.cpp
@@ -114,7 +114,7 @@ CoinMessages::CoinMessages(const CoinMessages & rhs)
   } else {
     char * temp = CoinCopyOfArray((char *) rhs.message_,lengthMessages_);
     message_ = (CoinOneMessage **) temp;
-    int offset = temp - (char *) rhs.message_;
+    long int offset = temp - (char *) rhs.message_;
     int i;
     //printf("new address %x(%x), rhs %x - length %d\n",message_,temp,rhs.message_,lengthMessages_);
     for (i=0;i<numberMessages_;i++) {
@@ -159,7 +159,7 @@ CoinMessages::operator=(const CoinMessages & rhs)
     } else {
       char * temp = CoinCopyOfArray((char *) rhs.message_,lengthMessages_);
       message_ = (CoinOneMessage **) temp;
-      int offset = temp - (char *) rhs.message_;
+      long int offset = temp - (char *) rhs.message_;
       int i;
       //printf("new address %x(%x), rhs %x - length %d\n",message_,temp,rhs.message_,lengthMessages_);
       for (i=0;i<numberMessages_;i++) {
@@ -479,7 +479,7 @@ CoinMessageHandler::gutsOfCopy(const CoinMessageHandler& rhs)
   longValue_=rhs.longValue_;
   charValue_=rhs.charValue_;
   stringValue_=rhs.stringValue_;
-  int offset ;
+  long int offset ;
   if (rhs.format_)
   { offset = rhs.format_ - rhs.currentMessage_.message();
     format_ = currentMessage_.message()+offset; }
